feat(example): California Housing XGBoost end-to-end UniFlow workflow
Replaces the Boston Housing XGBoost example with California Housing.
Boston Housing was removed by scikit-learn v1.2 over racial-bias concerns
(the engineered B feature). California Housing is a clean drop-in replacement.

The example is restructured from a single monolithic script into per-step
modules orchestrated by a @uniflow.workflow entry point:

  feature_prep.py   — Ray: load sklearn California Housing, train/test split
  preprocess.py     — Spark: cast columns to float
  train.py          — Ray: distributed XGBoost via XGBoostTrainer
  push_model.py     — Ray: push model + eval report via pusher API

Additional changes:
- MinioStorageBackend / LocalStorageBackend selected by MINIO_* env vars
- APIRegistryClient / InMemoryRegistryClient selected by REGISTRY_ENDPOINT
- All stateful objects constructed inside task body (UniFlow codec boundary)
- cache_enabled=True on feature_prep and preprocess
- README: pipeline diagram, UniFlow concepts, DatasetVariable explanation,
  local/remote run with --environ flags, breakpoint debugging, after-push
  serving guide (KServe/Seldon/BentoML)

Fix DatasetVariable.__init__ to accept _io_metadata so the UniFlow codec
can round-trip DatasetVariable instances across the workflow→task boundary.
Variable (the @dataclass parent) defines _io_metadata as a field, so
dataclass_dict() includes it; without this parameter the reconstruction
raises TypeError.
