+++ python/examples/california_housing_xgb/__init__.py
@@ -1,0 +1,5 @@
+"""California Housing XGBoost regression example."""
+
+from examples.california_housing_xgb.california_housing_xgb import train_workflow
+
+__all__ = ["train_workflow"]
